extract Viewer browser persistence module
